Move envs to own file. Err, not panic.
